rm Apache config file .htaccess
because the main use of it was for redirects that upgrade HTTP to
HTTPS, which we can accomplish using other means on the current
hosting service (NearlyFreeSpeech.net). As for other small
configuration details, they are not critical for the current site
design, but may need re-consideration if site begins to use CDN or
interacts with backend services.

[ci skip]